refactor(Crosshair): full refactor for modularity and easier extension
Complete refactor of Crosshair module:
- Improved code readability and organization
- Modularized modes for easier extension
- Streamlined event handlers and color management
