+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/crosshair/CrosshairMode.kt
@@ -1,0 +1,50 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2026 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.ccbluex.liquidbounce.features.module.modules.render.crosshair
+
+import net.ccbluex.liquidbounce.config.types.nesting.Choice
+import net.ccbluex.liquidbounce.config.types.nesting.ChoiceConfigurable
+import net.ccbluex.liquidbounce.event.events.OverlayRenderEvent
+import net.ccbluex.liquidbounce.event.handler
+import net.ccbluex.liquidbounce.features.module.modules.render.crosshair.ModuleCrosshair.modes
+import net.ccbluex.liquidbounce.utils.inventory.isInContainerScreen
+import net.ccbluex.liquidbounce.utils.inventory.isInInventoryScreen
+
+abstract class CrosshairMode(name: String) : Choice(name) {
+    final override val parent: ChoiceConfigurable<*>
+        get() = modes
+
+    protected val showInThirdPerson by boolean("ShowInThirdPerson", true)
+    protected val radius by intRange("Range", 3..5, 1..25)
+
+    protected abstract fun OverlayRenderEvent.drawCrosshair(centerWidth: Float, centerHeight: Float)
+
+    @Suppress("unused")
+    private val cursorHandler = handler<OverlayRenderEvent> {
+        if(!mc.options.cameraType.isFirstPerson && !showInThirdPerson) return@handler
+        if(isInInventoryScreen || isInContainerScreen) return@handler
+
+        val centerWidth = (it.context.guiWidth() / 2.002f)
+        val centerHeight = (it.context.guiHeight() / 2.0025f)
+
+        it.drawCrosshair(centerWidth, centerHeight)
+
+    }
+}
